+++ packages/at_client/lib/src/secret_sharing/secret_sharing.dart
@@ -25,6 +25,7 @@ export 'package:at_client/src/secret_sharing/algo_ids.dart';
 export 'package:at_client/src/secret_sharing/at_client_secret_sharing.dart';
 export 'package:at_client/src/secret_sharing/enrollment_directory.dart';
 export 'package:at_client/src/secret_sharing/enrollment_key_package.dart';
+export 'package:at_client/src/secret_sharing/enrollment_symmetric_key.dart';
 export 'package:at_client/src/secret_sharing/key_package.dart';
 export 'package:at_client/src/secret_sharing/key_package_registration.dart';
 export 'package:at_client/src/secret_sharing/pairwise_secret_sharing.dart';